backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
redis==5.0.1
Pillow==10.1.0
psycopg2-binary==2.9.9

